arabic-rag-chatbot/requirements.txt
# Core LangChain & LLM
langchain==0.1.14
langchain-openai==0.1.0
langchain-community==0.0.36

# Vector Database & Embeddings
qdrant-client>=1.8.0
langchain-qdrant>=0.1.0
langchain-cohere>=0.1.0
openai>=1.14.0

# Document Processing
pypdf==4.0.1
python-docx==0.8.11
python-pptx==0.6.21
unstructured

# Text Splitters (standalone, no deprecation warnings)
langchain-text-splitters>=0.0.1

# LangChain Core (canonical imports for zero-deprecation)
langchain-core>=0.1.0

# Database & Framework
sqlalchemy==2.0.25
asyncpg==1.0.0
alembic==1.13.1
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
cors==1.0.1

# Streamlit Frontend
streamlit==1.35.0
streamlit-chat==0.1.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0

# Logging & Monitoring & Testing
python-json-logger==2.0.7
sentry-sdk==1.38.0
